refactor: move packer/unpacker to internal
