+++ internal/pkg/image/packer/squashfs.go
