+++ internal/pkg/image/packer/squashfs_test.go
