+++ internal/pkg/image/unpacker/squashfs.go
