+++ internal/pkg/image/unpacker/squashfs_no_singularity.go
