+++ internal/pkg/image/unpacker/squashfs_singularity.go
