+++ pkg/image/image_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/sylabs/singularity/internal/pkg/util/fs"
 
 	imageSpecs "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/sylabs/singularity/pkg/image/unpacker"
+	"github.com/sylabs/singularity/internal/pkg/image/unpacker"
 )
 
 // We need a busybox SIF for these tests. We used to download it each time, but we have one
